updates to deploy to heroku

+++ src/app/front/front.component.ts
@@ -55,7 +55,7 @@ window.crypto.getRandomValues(challenge);
 
   str2ab(str) {
     let buf = new ArrayBuffer(str.length * 2); // 2 bytes for each char
-    let bufView = new Uint16Array(buf);
+    let bufView = new Uint8Array(buf);
     for (let i = 0, strLen = str.length; i < strLen; i++) {
       bufView[i] = str.charCodeAt(i);
     }
